Split entities package into characters, pickups and weapons.
