+++ core/src/main/java/com/elite/game/ClientGame.java
@@ -8,7 +8,11 @@ import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.math.Vector3;
 import com.badlogic.gdx.physics.box2d.*;
 import com.elite.audio.AudioAccessor;
-import com.elite.entities.*;
+import com.elite.entities.characters.Zombie;
+import com.elite.entities.pickups.AmmoCrate;
+import com.elite.entities.pickups.Crate;
+import com.elite.entities.pickups.HealthCrate;
+import com.elite.entities.weapons.Bullet;
 import com.elite.network.Client;
 import com.elite.ui.EnergyBar;
 import com.elite.ui.ServerListing;
