+++ core/src/main/java/com/elite/game/GameType.java
@@ -1,6 +1,6 @@
 package com.elite.game;
 
-import com.elite.entities.Zombie;
+import com.elite.entities.characters.Zombie;
 
 import java.util.ArrayList;
